refactor: resolve `global` package dependencies
* move MVC converters to each domain package

* move exceptions to each domain package
