+++ backend/demo/src/main/java/com/_2cha/demo/auth/exception/NoPrivilegeException.java
@@ -1,4 +1,6 @@
-package com._2cha.demo.global.exception;
+package com._2cha.demo.auth.exception;
+
+import com._2cha.demo.global.exception.ForbiddenException;
 
 public class NoPrivilegeException extends
                                   ForbiddenException {
