+++ backend/demo/src/main/java/com/_2cha/demo/review/exception/NoSuchReviewException.java
@@ -1,4 +1,6 @@
-package com._2cha.demo.global.exception;
+package com._2cha.demo.review.exception;
+
+import com._2cha.demo.global.exception.NotFoundException;
 
 public class NoSuchReviewException extends NotFoundException {
